refactor: (#53) 댓글 삭제 메서드 내 불필요한 변수 제거

+++ src/comments/comments.service.ts
@@ -90,11 +90,7 @@ export class CommentsService implements ICommentsService {
     postId: number,
   ): Promise<void> {
     await this.verifyPostAndGroup(postId, groupId);
-    const deletedComment = await this.verifyCommentOwnership(
-      id,
-      userId,
-      postId,
-    );
+    await this.verifyCommentOwnership(id, userId, postId);
     await this.commentsRepo.deleteComment(id);
   }
